feat(perf): geometric speed ramping with float speeds and naming cleanup
Replace hardcoded x2/÷2/÷4 speed jumps with configurable geometric
ramp_factor (default 1.5), expand speed range to 0.0625–16x with float
precision, and rename ambiguous target_speed/target_speed_multiplier to
max_speed/applied_speed throughout the stack.

+++ services/simulation/src/api/models/simulation.py
@@ -4,11 +4,11 @@ from pydantic import BaseModel
 
 
 class SpeedChangeRequest(BaseModel):
-    multiplier: int
+    multiplier: float
 
 
 class SpeedChangeResponse(BaseModel):
-    speed: int
+    speed: float
 
 
 class ControlResponse(BaseModel):
@@ -18,7 +18,7 @@ class ControlResponse(BaseModel):
 
 class SimulationStatusResponse(BaseModel):
     state: Literal["stopped", "running", "draining", "paused"]
-    speed_multiplier: int
+    speed_multiplier: float
     current_time: str
     # Detailed driver metrics
     drivers_total: int
